pyproject.toml
[project]
name = "gpt-bot"
version = "0.1.0"
description = "GPT bot は ChatGPT を Discord 上で利用するための Discord Bot です．"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "charset-normalizer>=3.4.2",
    "discord-py>=2.5.2",
    "docx2txt>=0.9",
    "duckduckgo-search>=8.0.2",
    "langchain-community>=0.3.24",
    "langchain-core>=0.3.61",
    "langchain-google-community>=2.0.7",
    "langchain-mcp-adapters>=0.1.1",
    "langchain[anthropic,google-genai,openai]>=0.3.25",
    "langgraph>=0.4.5",
    "openai>=1.82.0",
    "pymupdf>=1.26.0",
    "python-dotenv>=1.1.0",
    "rizaio>=0.11.0",
    "sympy>=1.14.0",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
]
